flask_app/requirements.txt
flask-sqlalchemy
simplejson
mysql-connector
flask_wtf
Flask-Less
repackage
redis